+++ webrtc/video_engine/vie_codec_impl.cc
@@ -653,7 +653,8 @@ bool ViECodecImpl::CodecValid(const VideoCodec& video_codec) {
   }
 
   if (video_codec.plType == 0 || video_codec.plType > 127) {
-    LOG(LS_ERROR) << "Invalif payload type: " << video_codec.plType;
+    LOG(LS_ERROR) << "Invalif payload type: "
+                  << static_cast<int>(video_codec.plType);
     return false;
   }
